refactored

+++ LinAlg/LinAlg/Headers/Matrix.h
@@ -18,14 +18,14 @@ public:
     Matrix operator*(const Matrix&);
     Matrix operator*(const Vector&);
 
-    void translate ( int, int, int);
-    void scale ( int, int, int);
+    void translate (double, double, double);
+    void scale (double, double, double);
     void rotate(double);
-    int& operator()(int row, int column);
-    int operator()(int row, int column) const;
+    double& operator()(int row, int column);
+    double operator()(int row, int column) const;
 
     void draw();
-    const std::vector<int> getData() const;
+    const std::vector<double> getData() const;
 
 private:
     int getIndex(int, int) const;
@@ -34,9 +34,9 @@ private:
 protected:
     bool hasMultiplicableDimension(const Matrix&);
     bool hasSameDimensions(const Matrix&);
-    int VectorInproduct(std::vector<int>, std::vector<int>);
+    int VectorInproduct(std::vector<double>, std::vector<double>);
 
-    std::vector<int> _data;
+    std::vector<double> _data;
     int _columnCount;
     int _rowCount;
 };
